move items logic in storage.service
