+++ src/components/FileExplorer/FileExplorer.tsx
@@ -63,7 +63,6 @@ interface FileExplorerState {
   fileInputRef: React.RefObject<HTMLInputElement>;
   email: string;
   token: string;
-  isDragging: boolean;
   isAdmin: boolean;
   isMember: boolean;
 }
@@ -76,13 +75,10 @@ class FileExplorer extends Component<FileExplorerProps, FileExplorerState> {
       fileInputRef: createRef(),
       email: '',
       token: '',
-      isDragging: false,
       isAdmin: true,
       isMember: false
     };
   }
-
-  moveEvent = {};
 
   get hasAnyItemSelected(): boolean {
     return this.props.selectedItems.length > 0;
@@ -192,128 +188,6 @@ class FileExplorer extends Component<FileExplorerProps, FileExplorerState> {
         reject(err);
       });
     });
-  }
-
-  openFolder = (e): Promise<void> => {
-    return new Promise((resolve) => {
-      this.props.dispatch(
-        storageThunks.fetchFolderContentThunk(e)
-      );
-      resolve();
-    });
-  };
-
-  updateNamesPaths = (user, contentFolders, namePath) => {
-    if (namePath.length === 0 || namePath[namePath.length - 1].id !== contentFolders.id) {
-      return user.root_folder_id === contentFolders.id ? 'All Files' : contentFolders.name;
-    }
-  }
-
-  clearMoveOpEvent = (moveOpId) => {
-    delete this.moveEvent[moveOpId];
-  };
-
-  decreaseMoveOpEventCounters = (isError, moveOpId) => {
-    this.moveEvent[moveOpId].errors += isError;
-    this.moveEvent[moveOpId].total -= 1;
-    this.moveEvent[moveOpId].resolved += 1;
-  };
-
-  move = (items: DriveItemData[], destination: number, moveOpId) => {
-    const isTeam = this.props.workspace === Workspace.Business;
-
-    // Don't want to request this...
-    if (items.some((item) => item.isFolder && item.id === destination)) {
-      return toast.warn('You can\'t move a folder inside itself\'');
-    }
-
-    // Init default operation properties
-    if (!this.moveEvent[moveOpId]) {
-      this.moveEvent[moveOpId] = {
-        total: 0,
-        errors: 0,
-        resolved: 0,
-        itemsLength: items.length
-      };
-    }
-
-    // Increment operation property values
-    this.moveEvent[moveOpId].total += items.length;
-
-    // Move Request body
-    const data = { destination };
-
-    let keyOp; // Folder or File
-    // Fetch for each first levels items
-
-    items.forEach((item) => {
-      keyOp = item.isFolder ? 'Folder' : 'File';
-      data[keyOp.toLowerCase() + 'Id'] = item.fileId || item.id;
-      fetch(`/api/storage/move${keyOp}`, {
-        method: 'post',
-        headers: getHeaders(true, true, isTeam),
-        body: JSON.stringify(data)
-      }).then(async (res) => {
-        const response = await res.json();
-        const success = res.status === 200;
-        const moveEvent = this.moveEvent[moveOpId];
-
-        // Decreasing counters...
-        this.decreaseMoveOpEventCounters(!success, moveOpId);
-
-        if (!success) {
-          toast.warn(`Error moving ${keyOp.toLowerCase()} '${response.item.name}`);
-        } else {
-          analyticsService.trackMoveItem(keyOp, {
-            file_id: response.item.id,
-            email: this.props.user.email,
-            platform: DevicePlatform.Web
-          });
-          // Remove myself
-          const items = this.props.items.filter((commanderItem: any) =>
-            item.isFolder
-              ? !commanderItem.isFolder ||
-              (commanderItem.isFolder && !(commanderItem.id === item.id))
-              : commanderItem.isFolder ||
-              (!commanderItem.isFolder && !(commanderItem.fileId === item.fileId))
-          );
-          // update state for updating commander items list
-
-          this.props.dispatch(storageActions.setItems(items));
-        }
-
-        if (moveEvent.total === 0) {
-          this.clearMoveOpEvent(moveOpId);
-          // If empty folder list move back
-          if (!this.props.items.length) {
-            this.folderTraverseUp();
-          }
-        }
-      });
-    });
-  };
-
-  removeFileFromFileExplorer = (filename) => {
-    const index = this.props.items.findIndex((item: DriveItemData) => item.name === filename);
-
-    if (!~index) {
-      // prevent undesired removals
-      return;
-    }
-
-    this.props.dispatch(
-      storageActions.setItems(Array.from(this.props.items).splice(index, 1))
-    );
-  }
-
-  folderTraverseUp() {
-    const { dispatch, namePath } = this.props;
-
-    // dispatch(storageActions.popNamePath);
-
-    dispatch(
-      storageThunks.fetchFolderContentThunk(_.last(namePath)?.id)
-    );
   }
 
   render(): ReactNode {
